refactor(server): asr cache to asr factory
