+++ apps/server/api/src/asr/mod.rs
@@ -1,19 +1,19 @@
-pub mod asr_cache;
 pub mod whisper;
 
-use crate::common::{ModelError, device};
-use candle_core::{Device, Tensor};
-use candle_nn::VarBuilder;
-use candle_transformers::models::whisper::{self as m, Config, audio};
-use tokenizers::Tokenizer;
-use whisper::{Model, Task, decoder::Decoder, multilingual::detect_language};
+use crate::common::ModelError;
+use crate::config;
+use async_trait::async_trait;
+use std::sync::{Arc, OnceLock};
+use tokio::sync::Mutex;
+use whisper::AsrWhisper;
 
+#[async_trait]
 pub trait Asr: Send + Sync {
-    fn transcribe(
+    async fn transcribe(
         &mut self,
         sample_rate: u32,
         samples: &[f32],
-    ) -> impl Future<Output = Result<RecognizerResult, ModelError>> + Send;
+    ) -> Result<RecognizerResult, ModelError>;
 }
 
 #[derive(Debug, Clone)]
@@ -23,120 +23,39 @@ pub struct RecognizerResult {
     pub prob: f32,
 }
 
-#[derive(Clone)]
-pub struct AsrWhisper {
-    device: Device,
-    model: Model,
-    tokenizer: Tokenizer,
-    config: Config,
-    mel_filters: Vec<f32>,
+static INSTANCE: OnceLock<AsrFactory> = OnceLock::new();
+
+pub struct AsrFactory {
+    default_instance: Arc<Mutex<Box<dyn Asr>>>,
 }
 
-impl AsrWhisper {
-    pub fn new(
-        model_path: String,
-        config_path: String,
-        tokenizer_path: String,
-    ) -> Result<Self, ModelError> {
-        let start = std::time::Instant::now();
-        let device = device(false)?;
-        let config: Config =
-            serde_json::from_str(&std::fs::read_to_string(config_path.clone()).map_err(|_e| {
-                ModelError::ModelFileNotFound(format!(
-                    "config file not found path = {}",
-                    config_path.clone()
-                ))
-            })?)
-            .map_err(|_e| {
-                ModelError::ModelInitFailure(format!(
-                    "file convert json failure path = {}",
-                    config_path
-                ))
-            })?;
-        let tokenizer = Tokenizer::from_file(tokenizer_path.clone()).map_err(|_e| {
-            ModelError::TokenFileNotFound(format!(
-                "token file not found,path = {} ",
-                tokenizer_path.clone()
-            ))
-        })?;
-        let mel_bytes = include_bytes!("melfilters128.bytes").as_slice();
-        let mut mel_filters = vec![0f32; mel_bytes.len() / 4];
-        <byteorder::LittleEndian as byteorder::ByteOrder>::read_f32_into(
-            mel_bytes,
-            &mut mel_filters,
-        );
-        let vb = unsafe { VarBuilder::from_mmaped_safetensors(&[model_path], m::DTYPE, &device)? };
-        let model = Model::Normal(m::model::Whisper::load(&vb, config.clone())?);
-        tracing::info!("loaded the model in {:?}", start.elapsed());
-        tracing::info!("model built");
-        Ok(Self {
-            device: device.clone(),
-            model,
-            tokenizer,
-            config,
-            mel_filters,
-        })
+impl AsrFactory {
+    pub fn new(default_instance: Arc<Mutex<Box<dyn Asr>>>) -> Self {
+        Self { default_instance }
     }
-}
 
-impl Asr for AsrWhisper {
-    async fn transcribe(
-        &mut self,
-        _sample_rate: u32,
-        samples: &[f32],
-    ) -> Result<RecognizerResult, ModelError> {
-        let device = &self.device;
-        let mel_filters = &self.mel_filters;
-        let config = &self.config;
-        let model = &mut self.model;
-        let tokenizer = &self.tokenizer;
+    pub async fn init() -> &'static Self {
+        INSTANCE.get_or_init(|| -> Self { Self::new(Arc::new(Mutex::new(Self::create_model()))) })
+    }
 
-        // TODO: setting
-        let seed: u64 = 299792458;
-        let task = Some(Task::Transcribe);
-        let timestamps = false;
-        let verbose = false;
+    pub fn global() -> &'static AsrFactory {
+        INSTANCE.get().unwrap()
+    }
 
-        let mel = audio::pcm_to_mel(config, samples, mel_filters);
-        let mel_len = mel.len();
-        let mel = Tensor::from_vec(
-            mel,
-            (1, config.num_mel_bins, mel_len / config.num_mel_bins),
-            device,
-        )?;
-        tracing::info!("loaded mel: {:?}", mel.dims());
+    pub fn default(&self) -> Arc<Mutex<Box<dyn Asr>>> {
+        self.default_instance.clone()
+    }
 
-        let language_token = Some(detect_language(model, tokenizer, &mel)?);
-
-        match Decoder::new(
-            model,
-            tokenizer,
-            seed,
-            device,
-            Some(language_token.clone().unwrap().0),
-            task,
-            timestamps,
-            verbose,
-        ) {
-            Ok(mut dc) => {
-                let result = dc.run(&mel);
-                tracing::info!("result = {:?}", result);
-                match result {
-                    Ok(result) => {
-                        let text = result
-                            .into_iter()
-                            .map(|item| item.dr.text)
-                            .collect::<String>();
-                        Ok(RecognizerResult {
-                            text,
-                            language: language_token.clone().unwrap().1.clone(),
-                            prob: language_token.clone().unwrap().2,
-                        })
-                    }
-                    Err(e) => Err(ModelError::Decoder(format!("decoder run error = {}", e))),
-                }
-            }
-            Err(e) => Err(ModelError::Decoder(format!("decoder new error = {}", e))),
-        }
+    pub fn create_model() -> Box<dyn Asr> {
+        let app_config = config::get();
+        let config = app_config.asr();
+        Box::new(
+            AsrWhisper::new(
+                config.model().to_string(),
+                config.config().to_string(),
+                config.tokens().to_string(),
+            )
+            .unwrap(),
+        )
     }
 }
